Complete Exercise 17.6: Time.int_to_time
I didn't complete this exactly as described by the book. I decided to
instead modify the `Time.int_to_time` implementation I had written in
Ch16 and not the one provided in the code sample. To do this I had to
convert the function to a method and also patch a mistake in the
unittest `setUp` method. The behaviour now seems to work as expected.