Match markov start & end of text to emma.txt
The author has hardcoded the pattern to match the start & end of both
the Gutenburg header and the text itself which must have been different
in the version of the book used because it caused the entire text to be
skipped when run agasint the version of Emma in this repository. For
simplicity the hardcoded flags have been replaced with indicators found
in the copy of Emma in this repository.